pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "distiller"
version = "0.1.5"

authors = [
  {name = "PamirAI Inc"},
  {email = "founders@pamir.ai"},
] 

readme = "README.md"
license = {file = "LICENSE"}

dependencies = [
    # images and math
    'Pillow==9.5.0',
    'numpy',
    'numba',
    # wifi / api
    'psutil',
    'requests'
]
requires-python = ">= 3.9"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
]

[project.optional-dependencies]
hardware = [
    # serials
    'spidev',
    'RPi.GPIO',
    'pyserial',
    # audio
    'pyaudio',
    'pyalsaaudio',
    # cam
    'rpi-libcamera',
    'rpi-kms',
    'picamera2'
]

